Restore PdfViewer host object COM visibility
